requirements.txt
discord.py>=2.0
aiohttp>=3.8
python-dotenv>=1.0
aiosqlite>=0.19
